Use Uint16Array

+++ loader/ALLoader.js
@@ -432,7 +432,7 @@ export class ALLoader extends THREE.Loader {
 		for (const oneArray of multiArrays) {
 			bufferSize += oneArray.length;
 		}
-		const buffer = toTypedArray ? (toFloat32Array ? new Float32Array(bufferSize) : new Uint32Array(bufferSize)) : new Array(bufferSize);
+		const buffer = toTypedArray ? (toFloat32Array ? new Float32Array(bufferSize) : new Uint16Array(bufferSize)) : new Array(bufferSize);
 		let index = 0;
 		let materialIndex = 0;
 		for (const oneArray of multiArrays) {
@@ -454,7 +454,7 @@ export class ALLoader extends THREE.Loader {
 			bufferSize += indices.length;
 		}
 		bufferSize *= itemSize;
-		const buffer = toFloat32Array ? new Float32Array(bufferSize) : Uint32Array(bufferSize);
+		const buffer = toFloat32Array ? new Float32Array(bufferSize) : Uint16Array(bufferSize);
 		let index = 0;
 		for (const indices of multiIndices) {
 			for (let i = 0; i < indices.length; i++) {
